Rewritten tai library for negative durations

+++ src/tai/caldate_mjd.c
@@ -1,8 +1,8 @@
 #include "caldate.h"
 
-static unsigned long times365[4] = { 0, 365, 730, 1095 } ;
-static unsigned long times36524[4] = { 0, 36524UL, 73048UL, 109572UL } ;
-static unsigned long montab[12] =
+static uint32_t times365[4] = { 0, 365, 730, 1095 } ;
+static uint32_t times36524[4] = { 0, 36524UL, 73048UL, 109572UL } ;
+static uint32_t montab[12] =
 { 0, 31, 61, 92, 122, 153, 184, 214, 245, 275, 306, 337 } ;
 /* month length after february is (306 * m + 5) / 10 */
