+++ src/tai/taia_pack.c
@@ -4,7 +4,7 @@ void taia_pack(s,t)
 char *s;
 struct taia *t;
 {
-  unsigned long x;
+  uint32_t x;
 
   tai_pack(s,&t->sec);
   s += 8;
