+++ src/tai/taia_sub.c
@@ -7,9 +7,9 @@ struct taia *t;
 struct taia *u;
 struct taia *v;
 {
-  unsigned long unano = u->nano;
-  unsigned long uatto = u->atto;
-  
+  uint32_t unano = u->nano;
+  uint32_t uatto = u->atto;
+
   t->sec.x = u->sec.x - v->sec.x;
   t->nano = unano - v->nano;
   t->atto = uatto - v->atto;
